+++ hadoop/namenod_rpcactivity8020.go
@@ -36,7 +36,7 @@ type RPCActivityForPort8020Bean struct {
 	RPCProcessingTimeNumOps          *int64   `json:"RpcProcessingTimeNumOps,omitempty"`
 	RPCProcessingTimeAvgTime         *float64 `json:"RpcProcessingTimeAvgTime,omitempty"`
 	DeferredRPCProcessingTimeNumOps  *int64   `json:"DeferredRpcProcessingTimeNumOps,omitempty"`
-	DeferredRPCProcessingTimeAvgTime *int64   `json:"DeferredRpcProcessingTimeAvgTime,omitempty"`
+	DeferredRPCProcessingTimeAvgTime *float32 `json:"DeferredRpcProcessingTimeAvgTime,omitempty"`
 	RPCAuthenticationFailures        *int64   `json:"RpcAuthenticationFailures,omitempty"`
 	RPCAuthenticationSuccesses       *int64   `json:"RpcAuthenticationSuccesses,omitempty"`
 	RPCAuthorizationFailures         *int64   `json:"RpcAuthorizationFailures,omitempty"`
