+++ hadoop/resourcemanager_clusterinfo.go
@@ -23,20 +23,20 @@ type ClusterMetrics struct {
 }
 
 type ClusterMetricBean struct {
-	Name                   *string `json:"name,omitempty"`
-	ModelerType            *string `json:"modelerType,omitempty"`
-	TagClusterMetrics      *string `json:"tag.ClusterMetrics,omitempty"`
-	TagContext             *string `json:"tag.Context,omitempty"`
-	TagHostname            *string `json:"tag.Hostname,omitempty"`
-	NumActiveNMS           *int64  `json:"NumActiveNMs,omitempty"`
-	NumDecommissioningNMS  *int64  `json:"NumDecommissioningNMs,omitempty"`
-	NumDecommissionedNMS   *int64  `json:"NumDecommissionedNMs,omitempty"`
-	NumLostNMS             *int64  `json:"NumLostNMs,omitempty"`
-	NumUnhealthyNMS        *int64  `json:"NumUnhealthyNMs,omitempty"`
-	NumRebootedNMS         *int64  `json:"NumRebootedNMs,omitempty"`
-	NumShutdownNMS         *int64  `json:"NumShutdownNMs,omitempty"`
-	AMLaunchDelayNumOps    *int64  `json:"AMLaunchDelayNumOps,omitempty"`
-	AMLaunchDelayAvgTime   *int64  `json:"AMLaunchDelayAvgTime,omitempty"`
-	AMRegisterDelayNumOps  *int64  `json:"AMRegisterDelayNumOps,omitempty"`
-	AMRegisterDelayAvgTime *int64  `json:"AMRegisterDelayAvgTime,omitempty"`
+	Name                   *string  `json:"name,omitempty"`
+	ModelerType            *string  `json:"modelerType,omitempty"`
+	TagClusterMetrics      *string  `json:"tag.ClusterMetrics,omitempty"`
+	TagContext             *string  `json:"tag.Context,omitempty"`
+	TagHostname            *string  `json:"tag.Hostname,omitempty"`
+	NumActiveNMS           *int64   `json:"NumActiveNMs,omitempty"`
+	NumDecommissioningNMS  *int64   `json:"NumDecommissioningNMs,omitempty"`
+	NumDecommissionedNMS   *int64   `json:"NumDecommissionedNMs,omitempty"`
+	NumLostNMS             *int64   `json:"NumLostNMs,omitempty"`
+	NumUnhealthyNMS        *int64   `json:"NumUnhealthyNMs,omitempty"`
+	NumRebootedNMS         *int64   `json:"NumRebootedNMs,omitempty"`
+	NumShutdownNMS         *int64   `json:"NumShutdownNMs,omitempty"`
+	AMLaunchDelayNumOps    *int64   `json:"AMLaunchDelayNumOps,omitempty"`
+	AMLaunchDelayAvgTime   *float32 `json:"AMLaunchDelayAvgTime,omitempty"`
+	AMRegisterDelayNumOps  *int64   `json:"AMRegisterDelayNumOps,omitempty"`
+	AMRegisterDelayAvgTime *float32 `json:"AMRegisterDelayAvgTime,omitempty"`
 }
